chore: colocate unit tests
