+++ components/Input/Input.test.tsx
@@ -1,22 +1,12 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import Input from "@/components/Input/Input";
-import Select from "@/components/Select/Select";
-import Textarea from "@/components/Textarea/Textarea";
 
-test("text inputs link labels and errors", async () => {
+test("input links label and error", async () => {
     render(
         <form>
             <label htmlFor="name">Name</label>
             <Input id="name" data-testid="input" />
-
-            <label htmlFor="bio">Bio</label>
-            <Textarea id="bio" data-testid="textarea" />
-
-            <label htmlFor="choice">Choice</label>
-            <Select id="choice" data-testid="select">
-                <option value="1">One</option>
-            </Select>
 
             <label htmlFor="error">Error</label>
             <Input
